+++ src/main/java/hive/FlowerIntelligenceSink.java
@@ -1,0 +1,9 @@
+package hive;
+
+/**
+ * @author nick
+ * @since 2019/10/30
+ */
+public interface FlowerIntelligenceSink {
+    void foundFlowers(String workerBeeName);
+}